Fixed sprite facing due to wrong angular position assumption.
Added simple documentation to the getDirection() method and updated
component trigger info.

There was an issue visible in Demo 2. When mouse cursor was positioned
Left relative to sprite position, the sprite was facing Right instead of
Left. This issue occurred due to the fact that it was assumed the sprite
was in the center of the coordinate system when calculating the
relative angle where in fact it is the mouse cursor that holds the
central position.